Iterate leaders only to find sync lock groups; makes no difference...
... A sync lock group is a maximal subsequence of the track list containing
one or more wave and/or note tracks, followed by zero or more label tracks; or,
just one label track alone, which is not preceded by any wave or note tracks.

That's what was implemented; so skipping right wave track channels to find the
range of tracks makes no difference.