Call GetIntervals only on leaders, changes misaligned behavior...
... Snapping ignores clip boundaries in right channels

Also might be a difference adjusting the record-to-play change time for
punch-and-roll recording into a misaligned stereo track -- extremely unlikely
use case!

Might change snapping behavior when dragging a right channel clip affordance,
but those will be eliminated later.

This replaces two calls to TrackList::Find() with FindLeader(); the only
remaining calls now are implementation details in WaveTrack and TrackList