Replace Channel::GetOffset() everywhere with GetStartTime()...
... They differed only for LabelTracks, for which the offset was always 0 but
the start time was the start of the leftmost label (or 0 when there are no
labels).  (mOffset is reassigned only in ChannelGroup::MoveTo() but LabelTrack
overrides it.)

For Time track, offset and start time are always 0.

"J" key behavior may change when a label track is selected, also click on the
"Select" button of a track, ctrl + click in a track, etc.  The "start" of the
label track is now the start of its leftmost label, not always 0.

Some other uses of GetOffset() were only for wave tracks (Export.cpp) or MIDI
tracks.