Fix: Copy/Paste ...
The code that intended to restore splits is now gone. It was asserting when checking that some clip's right trim value was zero (places commented with "// precondition satisfied because... ??". I think this had dependencies to what happens in `HandleClear`. Anyway, I couldn't even get to bottom of what this was intending to do. Testing has shown that cut lines are preserved as before, though.