Pressed state of the "play" and "pause" buttons are not provided in Msaa
Issue #4991.
Problem:
In commit 67aad3504, the accessibility of AButtons when set to be toggle buttons was improved for all screen readers and for Narrator in particular. This involved both the role and state being included in the accessibility name. However, because the accessibility state no longer indicated the state of the toggle button, this made it difficult for screen reader scripts to robustly determine the state of the toggle button.

Fix:
Add the character '\a' to the end of the accessibility name when the toggle button is in the pressed state.