Get ChannelGroupData from holder and not owner
When clips are added to `TrackList::mPendingTracks` attemp to access `ChannelGroupData` of the (right)track results in a new instance of `ChannelGroupData` being created with default sample rate that could be different from the sample rate of the original track.
Not the case for the first track's copy as it already contains a copy of `ChannelGroupData` originated from the source track.