lib-realtime-effects uses lib-channel but not lib-mixer...
... and lib-mixer does not need lib-registries because WideSampleSequence does
not need to be a Site.

lib-audio-io gets lib-channel transitively but must depend on lib-mixer
directly.

RealtimeEffectList now stores the effect stack as a ChannelGroup attachment.

The ClientData::Site inside the ChannelGroup is the only part of ChannelGroup's
interface that it relies on.  But ChannelGroup is otherwise a very low-level
and small library.

Cloning of the realtime effect stack (as when splitting stereo to mono) now
comes for free with the rest of channel group data duplication and swapping of
stereo channels simply keeps the identical channel group data including the
stack.

WaveTrack's interface shrinks!  Reinit isnt' needed.