Fix bar formats when using a screen reader.
Bar formats in time controls have been changed so that the size of the first field can automatically increase. This enables bar numbers greater than 999.

Problem:
NumericTextCtrlAx::GetName() uses the position of the focused digit from the start of the digits to determine whether the focus has moved to a different digit. However this doesn't work if extra digits are added at the start of the first field, and can cause screen readers to read the wrong value as described in issue #5344.

Fix:
Use the position of the focused digit from the end of the digits to determine whether the focus has moved to a different digit.