Export dialog: keyboard navigation of radio buttons
Problem:
Since commit https://github.com/audacity/audacity/pull/5447, the multiple files radio button can be unavailable. When this is the case and both the entire project and then current selection radio buttons are available, if the entire project radio button is the focus and selected, then pressing down arrow does not change the selection, but moves the focus to the edit metadata button.

Fix:
When there are not multiple audio tracks and no labels, leave the multiple files radio button available.