Compressor: fix for when there is blank space before audio
Problem:
When the time selection includes blank space before the audio, the results of the compressor effect are incorrect.

Relevant previous commits:
618106f1
0f830b44
The latter commit is a fix for the former. In the fix, a period of silence, mT0, is inserted into the WorkTracks, so that in ProcessOne() samples are appended starting at the same time as the start of processed audio in the original track.
However, in ProcessPass(), the start of processing is given by std::max(trackStart, mT0), not mT0.

Fix:
Insert a period of silence based on std::max(trackStart, mT0).
In addition, because after the addition of the silence the WorkTracks start from 0, the value of trackStart isn't calculated using WorkTracks, so avoiding unnecessary processing. trackEnd is handled similarly, just for consistency, although this isn't necessary.