Music import dialog: screen reader problem
When the Music import dialog opens, it's not automatically read by NVDA or Narrator, but it is by Jaws.
Both NVDA and Narrator still think that the focus is the imported track.

Problem:
In TrackPanelAx::GetState(), the state of the list or a list item is given as focussed without checking whether it is actually focussed.
Looks like this has always been the case, but hasn't been a problem up to now.

Fix:
Check whether focused: For the list item, explicitly check, and the for list, let wxWidgets do this using a standard accessible object.