Timer Record: incorrect default number of channels.
Timer Record automatic export defaults to mono when project is set to record in stereo

Problem:
In the TimerRecordDialog class, m_iAutoExportChannels is initialized to 0;
In the TimerRecordDialog, if you enable automatic export, but don't press the Select button to open the Export Audio dialog, then this leads to the export failing.
If you do press the Select button and open the Export Audio dialog, then initially channels is set to mono, even if you're going to record in stereo.
(There's a a call to TimerRecordExportDialog::Bind with number of channels equal to 0, then a call to ExportFilePanel::Init, with number of channels 0.
When this parameter is zero this Init function gets the correct number of channels from the tracks to be exported, but in this context, there are as yet no tracks.)

Fix:
Set m_iAutoExportChannels to the number of recording channels.