Tempo and upper time signature: problem with some keyboards
Problem:
Tempo and upper time signature: using German or Norwegian keyboard, can't enter digits 8 or 9.

When the digits 8 or 9 are typed (or any digit) MenuCreator::FilterKeyEvent() returns these to wxWidgets for further processing, without them being matched to the shortcuts of any commands.
However, on Windows, on being returned to wxWidgets, any keys are first matched with accelerators in the menus, before being given to the control which is the focus. In the German and Norwegian keyboard layouts, [ and ] are AltGr 8 and AltGr 9. Not sure quite why, but 8 and 9 match the keyboard accelerators [ and ].

Fix:
In FormatLabelWithDisabledAccel(), add [ and ] to the accelerators which are disabled.