Playback cursor issues after seeking.
Problem:
After seeking, there are a number of cases in which the playback does not correspond to the playback position. A number of these are listed in the issue: https://github.com/audacity/audacity/issues/2133

The problem was caused by this commit: 5af2a26, which removed the constraints on seeking in PlaybackPolicy::OffsetTrackTime().

Fix:
Reinstate the constraints for seeking in PlaybackPolicy(), and add DefaultPlaybackPolicy::OffsetSequenceTime() with appropriate constraints for this policy.