Pitch and speed dialog: names read by screen readers are incorrect
Problem A:
The names of the first two controls are not read correctly by screen readers.
These two controls are wxSpinCtrl, which on Windows create a text box which is a "buddy window", and which is the focus. This makes it very difficult to set a WindowAccessible so that the accessibility name can be made to include text coming after the control.

Fix A:
Change the layout of the labels of the controls so that Windows can automatically set the accessibility names such that they are useful.

Problem B:
The accessibility name of the third control is just "%", rather than "Clip Speed %".

Fix A:
Just add "Clip Speed" to the accessibility name using a standard ShuttleGui function.