Import/Export preferences: accessibility issue
Problem:
The names of the group boxes are not read by Jaws or Narrator. But they are by NVDA.

The radio buttons are grandchildren of the group box, but they should be children.
Extra panels where introduced by this commit: 7a55c90, to fix a problem with keyboard navigation of radio buttons on macOS:
https://bugzilla.audacityteam.org/show_bug.cgi?id=2692
And I presume an extra panel was added in the Music imports group to follow this pattern.

Fix:
Only include the extra panel on Mac builds.