Move the Timeline sub menu
Problem:
Timeline sub menu isn't easy to find in the extra menu, which isn't shown by default, and the timeline view has increased importance in the upcoming 3.5. release.

Fix:
Move the timeline sub menu to the View menu.
The only available access key for Timeline on the View menu was the letter l. Given that is not a desirable access key due to the shortness of the underline indicating the access key, the access key "e" was chosen, and the access key of another item changed.