Export multiple confirmation dialog: make Enter close dialog
Problem:
In this dialog, a read-only text box showing the files exported is the initial focus. A user would expect that pressing Enter would close the dialog without the need to first tab to the OK button, but this is not the case.

Fix:
Use the wxEVT_TEXT_ENTER event, to close the dialog when the text box is the focus, and the user presses Enter.