Record new track can record at the wrong sampling rate
Problem:
If an audio track is selected, then when recording a new track, the actual sampling rate for the recording is the sampling rate of the selected track, rather than the sampling rate of the project. The sampling rate of the new track in the track control panel is the sampling rate of the project. So if these two sampling rates are different, this results in the recording with a changed speed.

Looks like the problem was introduced by this commit: 70b7487

Fix:
Only use the sampling rate of the selected track or tracks when appending to them.