Pitch and speed dialog: make it reachable via the Alt + F6 shortcut
Problem:
With the full default shortcuts enabled, pressing Alt + F6 or Shift Alt + F6 cycles you round the main window and any open modeless dialogs. However, the Pitch and speed dialog is not included in this cycle.
For this to be the case the dialog's parent should be the project window. (See in NavigationMenus.cpp, OnPrevWindow() and OnNextWindow().)
This isn't the case for the Pitch and speed dialog, the parent is nullptr.

Fix:
Change the parent of the Pitch and speed dialog to be the project window.