Compressor and Limiter accessibility: check boxes
Problem:
When the compressor and limiter are used as realtime effects, there are a number of check boxes immediately above a graph which determine which parameters are shown in the graph. For sighted users, the purpose of these check boxes is reasonably obvious.
However, for users of screen readers, the names of the check boxes, eg input, output, without their visual context, do not make their purpose at all obvious.

Fix:
Change the accessibility names of the check boxes to "Show input", "Show output", etc.