Disabled Settings menu item
On macOS, we can’t open preferences from the menu now - the preferences item now looks like a menu with a Sort items item.
This happened because we have another menu Extra->Settings. In macOS, this word is reserved too for calling preferences. So, we cannot add two Preferences and Settings items at the same time.