Compressor and limiter accessibility: using the non-linear sliders
Problem:
When using the exponential sliders with arrow keys, there are two issues:
A. Right/down arrow key does not change the position.

B. When the position is near the left hand end of the slider, the arrow keys don't change the position of the slider. The setting value has a precision of 1 decimal place. Near the left hand end of the slider, a change is slider position by 1, may result in a change setting value, when "rounded" to 1 decimal place of zero. So the slider position and setting value is unchanged.

Fixes:
A. In bool DynamicRangeProcessorEditor::UpdateUI(), round the value when setting the slider positions.

B. Handle the wxEVT_SCROLL_LINEDOWN and wxEVT_SCROLL_LINEUP events, and ensure that the change in the slider position results in a change of setting value of at least 0.1.