Usability of various sliders for keyboard users
A number of bugs were introduced by a PR for changing the play-at-speed accuracy back to 3 decimal places (commit: 37330d6).
These bugs are all fixed by reverting some of the unnecessary changes in that PR.

A
Problem: the step size for keyboard users of ASliders is too small (For example the recording and playback level sliders). This is fixed by reverting the constant 100.0 to 10.0 in LWSlider::Increase() and LWSLider::Decrease().

B
Problem: For the play-at-speed slider, using the keyboard, it's not easy to get back to a speed of exactly 1.000. This was caused by the change of the minimum of the speed slider being changed from 0.01 to 0.001, and the dependence of LWSlider::Increase() and LWSlider::Decrease() on that value. So the minimum value of the speed slider was reverted to 0.01 (which I think is quite slow enough).

C
Problem: The commands Extra > Play at speed > Increase/Decrease playback speed, result in very small changes of speed. Fixed by reverting the constant 0.01 to 0.1 in OnPlaySpeedInc() and OnPlaySpeedDec()