VST3 Effects: fix problem with discrete parameters in plain UI.
For some VST3 effects, for discrete parameters in the plain UI for keyboard users, it's not possible to set all the possible values.

It appears that the problem is that some plugins don't provide the correct implementation of the functions for converting between plain and normalized values.

Fix:
Detect when invalid functions for these conversions are provided, and when this is the case, use the formula use the formula given by Steinberg in their documentation.
(Form posts on Reaper's forums, it appears that Reaper does something similar to this.)