Importing audio files: do not change time selection
Problem:
In 3.6.0, when importing audio files, the time selection is set to cover the audio in the last imported file. This is a change wrt previous versions, can means that users loose any previously set time selection.
It was introduced by the commit: 4532f561db

Fix:
Remove the setting of the time selection when importing audio files.