improve misleading diagnostics on mod load failure
wxSysErrorMsg() uses strerror(errno) on Unix platforms.
If Load() failed due to something other than a system
call failure, this will be a misleading diagnostic (such
failures happen in practice due to libdl failures in
wxDynamicLibrary). Zero out errno before calling Load(),
so at worst we'll get "Error: Operation succeeded" or
whatever, so the confusion is at least tightly
constrained. In addition, remove wxDL_QUIET, so that
we get a diagnostic printed to stderr on a libdl failure;
this ought make the true problem obvious.

Signed-off-by: nick black <dankamongmen@gmail.com>