Application preferences: fix text not being read by screen readers
Problem:
In Application preferences, the descriptive text below the two check boxes is not automatically read by screen readers.

Fix:
Append the descriptive text to the accessibility names of the check boxes. This does make for very long accessibility names, which is not ideal, but at least the text is automatically read.