Welcome dialog: accessible names for carousel buttons.
Problem: the left and right arrow buttons have no accessible names, and the accessible name of the middle button is just its label. Users of screen readers need access to the information on the slides, and to be able to interact with the carousel.

Fix: Add "previous slide" and "next slide" accessible names to the left and right arrow buttons. Also add the slide title, and where helpful any text in the image to the accessibility name of the middle button.