Horizontal scrolling - fix crash of debug version on Windows
Problem:
When no horizontal scrolling is possible, trying to scroll by dragging using the middle mouse button causes a debug exception.

When this is the case, in the call to void Viewport::ScrollHorizontalByPixels(int deltaPixels),
in the call to std::clamp(current - deltaPixels, 0, max), max can be negative, leading to clamp() being undefined.

Additional issue, scrollbar thumb moves in the wrong direction.

Fix:
Ensure max is not negative (as is done in other funtions in the file, eg Viewport::SetHorizontalThumb()).

Scrollbar thumb now moves in the correct direction.